Move TriggerTargets into its own module
